backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
google-genai>=0.1.0
python-dotenv>=1.0.0
gunicorn>=23.0.0
groq>=0.5.0
